Added Dynamic Monster Health & Dynamic Mana Pool
- TOWERS
- BUTTONS
- FIREBALLS
- Game Over when ManaBar == 0
- Pause
- 2x Speed
- healthDecrease
- interaction of Buttons

+++ src/main/java/WizardTD/ManaBar.java
@@ -19,7 +19,7 @@ public class ManaBar {
         }
     }
 
-    public void decreaseMana(int amount) {
+    public void decreaseMana(float amount) {
         currMana -= amount;
         if (currMana < 0) {
             currMana = 0; // Ensure mana doesn't go negative
